requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.2.2
zipp==3.8.1
